Encoding shouldn't be a problem anymore:
Declare in the translate url the incoming and outcoming encoding (UTF-8) which makes it clear what to use to decode the incoming response.
Second problem that existed in the past. Non-ASCII chars submitted would end in failure. This is circumvented by urllib.parser.quote_plus which encodes the words to be useable in URLs.

This closes and  fixes #2 and #7.